pyproject.toml
[tool.poetry]
name = "pokemon-api"
version = "0.1.0"
description = ""
authors = ["Кирилл Запьянцев <kzapyantsev@rbc.ru>"]
readme = "README.md"
# packages = [{ include = "src/pokemon" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
motor = "^3.3.1"
python-dotenv = "^1.0.0"
structlog = "^23.1.0"
uvicorn = "^0.23.2"
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
pydantic-settings = "^2.0.3"
gunicorn = "^21.2.0"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[tool.isort]
profile = "black"
multi_line_output = 3
combine_as_imports = true
line_length = 88

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
